% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC.R
\name{MCMC}
\alias{MCMC}
\alias{log_likelihood_i}
\alias{log_likelihood}
\alias{log_prior_flat_exp}
\alias{log_prior}
\alias{log_posterior}
\alias{move_shape}
\alias{move_scale}
\title{Functions to run an MCMC to estimate the parameters of a gamma distribution}
\usage{
log_likelihood_i(i, dat, shape, scale)

log_likelihood(dat, shape, scale)

log_prior_flat_exp(param, mean_prior = 1000)

log_prior(shape, scale, mean_prior_shape = 1000, mean_prior_scale = 1000)

log_posterior(
  dat,
  shape,
  scale,
  mean_prior_shape = 1000,
  mean_prior_scale = 1000
)

move_shape(
  curr_val_shape,
  curr_val_scale,
  sdlog,
  dat,
  mean_prior_shape = 1000,
  mean_prior_scale = 1000
)

move_scale(
  curr_val_shape,
  curr_val_scale,
  sdlog,
  dat,
  mean_prior_shape = 1000,
  mean_prior_scale = 1000
)
}
\arguments{
\item{i}{stub}

\item{dat}{stub}

\item{shape}{stub}

\item{scale}{stub}

\item{param}{stub}

\item{mean_prior}{stub}

\item{mean_prior_shape}{stub}

\item{mean_prior_scale}{stub}

\item{curr_val_shape}{stub}

\item{curr_val_scale}{stub}

\item{sdlog}{stub}
}
\value{
stub
}
\description{
These function enable us to fit a gamma distribution to data
}
\details{
NOTE EVERYWHERE WE WORK ON THE LOG SCALE
}
\author{
Anne Cori & Joshua Lambert
}
\keyword{internal}
